Add files via upload
Google TextToSpeech and SpeechToText activities that allows you to give voice to your robots.
The attended robot is receiving all details through voice commands, by using this package.
A html form pop-up to allow the user to see all the details received through voice. 
As soon the robot has all the details is start searching on multiple sites for the best deals.
The offers are return into a custom html page in order to allow the user to choose.